requirement.txt
Click==7.0
Django==2.2.5
django-cors-headers==3.1.0
django-rest-auth==0.9.5
djangorestframework==3.10.3
dlib==19.17.0
face-recognition==1.2.3
face-recognition-models==0.3.0
Flask==1.1.1
imutils==0.5.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysqlclient==1.4.2
numpy==1.16.4
opencv-python==4.1.0.25
paho-mqtt==1.4.0
Pillow==6.1.0
PyMySQL==0.9.3
pytz==2019.2
scipy==1.3.1
six==1.12.0
SQLAlchemy==1.3.7
sqlparse==0.3.0
Werkzeug==0.15.6
